fixed reading a double for spin ctrl double
git-svn-id: http://svn.xp-dev.com/svn/wxextension@1918 f22100f3-aa73-48fe-a4fb-fd497bb32605

+++ trunk/extension/src/configdlg.cpp
@@ -562,7 +562,7 @@ wxControl* wxExConfigDialog::AddSpinCtrlDouble(wxWindow* parent,
     inc,
     m_Config->ReadLong(m_ConfigGroup + text, min));
 
-  spinctrl->SetValue(m_Config->ReadLong(m_ConfigGroup + text, min));
+  spinctrl->SetValue(m_Config->ReadDouble(m_ConfigGroup + text, min));
 
   return Add(sizer, parent, spinctrl, text + ":", false);
 }
